Fix coding standards issues

+++ web/modules/custom/event_pull/src/Model/Event.php
@@ -10,14 +10,14 @@ class Event {
   /**
    * The original event data.
    *
-   * @var \stdClass
+   * @var object
    */
   private $eventData;
 
   /**
    * Event constructor.
    *
-   * @param \stdClass $eventData
+   * @var object $eventData
    *   The event data.
    */
   public function __construct(\stdClass $eventData) {
@@ -25,16 +25,23 @@ class Event {
   }
 
   /**
-   * The name of the event.
+   * Get the name of the event.
    *
    * @return string
+   *   The event name.
    */
   public function getName(): string {
     return $this->eventData->name;
   }
 
+  /**
+   * Get the date of the event.
+   *
+   * @return int
+   *   The event date timestamp.
+   */
   public function getDate(): int {
-    return substr($this->eventData->time, 0, 10);
+    return (int) substr($this->eventData->time, 0, 10);
   }
 
 }
